Refactor into modules with IPC
