+++ .gitignore
@@ -1,0 +1,5 @@
+__pycache__/
+*.pyc
+*.pkl
+*.json
+
