+++ requirements.txt
@@ -1,0 +1,5 @@
+flet
+sentence-transformers
+scikit-learn
+numpy
+
